requirements.txt
Scrapy~=2.6.2
itemadapter~=0.7.0
google-cloud-bigquery>=3.3.2